# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [patients]
      summary: Request a private medical report upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlInput"
      responses:
        "200":
          description: Signed upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResult"
components:
...
  schemas:
...
    RequestUploadUrlInput:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string }
        size: { type: integer }
        contentType: { type: string }
    RequestUploadUrlResult:
      type: object
      required: [uploadURL, objectPath, metadata]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
        metadata:
          type: object
          required: [name, size, contentType]
          properties:
            name: { type: string }
            size: { type: integer }
            contentType: { type: string }