# lib/api-zod/src/generated/types/requestUploadUrlInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RequestUploadUrlInput {
  name: string;
  size: number;
  contentType: string;
}